SDI12 review
- We aren't using concurrent commands, so measure/data one sensor after
  the other

- CTD-10, wait for the number of seconds specified by the M command

- Check the returned value starts with the expected address

+++ libraries/SDI12/SDI12.h
@@ -99,7 +99,7 @@ public:
   uint8_t readline();
   uint8_t command2address(uint8_t address, const char* cmd);
   uint8_t identification(uint8_t address);
-  uint8_t measure(uint8_t address);
+  int measure(uint8_t address);
   uint8_t data(uint8_t address);
 
   char read_address();
